Adding Idempotency test
