+++ tests/functional/idempotency/_boto3/test_idempotency.py
@@ -8,7 +8,6 @@ import jmespath
 import pytest
 from botocore import stub
 from botocore.config import Config
-from pydantic import BaseModel
 from pytest import FixtureRequest
 from pytest_mock import MockerFixture
 
@@ -32,7 +31,6 @@ from aws_lambda_powertools.utilities.idempotency.exceptions import (
     IdempotencyInconsistentStateError,
     IdempotencyInvalidStatusError,
     IdempotencyKeyError,
-    IdempotencyModelTypeError,
     IdempotencyNoSerializationModelError,
     IdempotencyPersistenceLayerError,
     IdempotencyValidationError,
@@ -47,9 +45,6 @@ from aws_lambda_powertools.utilities.idempotency.serialization.custom_dict impor
 from aws_lambda_powertools.utilities.idempotency.serialization.dataclass import (
     DataclassSerializer,
 )
-from aws_lambda_powertools.utilities.idempotency.serialization.pydantic import (
-    PydanticSerializer,
-)
 from aws_lambda_powertools.utilities.validation import envelopes, validator
 from aws_lambda_powertools.warnings import PowertoolsUserWarning
 from tests.functional.idempotency.utils import (
@@ -61,7 +56,7 @@ from tests.functional.idempotency.utils import (
 from tests.functional.utils import json_serialize, load_event
 
 TABLE_NAME = "TEST_TABLE"
-TESTS_MODULE_PREFIX = "test-func.tests.functional.idempotency.test_idempotency"
+TESTS_MODULE_PREFIX = "test-func.tests.functional.idempotency._boto3.test_idempotency"
 
 
 def get_dataclasses_lib():
@@ -1316,106 +1311,6 @@ def test_idempotent_function_serialization_no_response():
 
 
 @pytest.mark.parametrize("output_serializer_type", ["explicit", "deduced"])
-def test_idempotent_function_serialization_pydantic(output_serializer_type: str):
-    # GIVEN
-    config = IdempotencyConfig(use_local_cache=True)
-    mock_event = {"customer_id": "fake", "transaction_id": "fake-id"}
-    idempotency_key = f"{TESTS_MODULE_PREFIX}.test_idempotent_function_serialization_pydantic.<locals>.collect_payment#{hash_idempotency_key(mock_event)}"  # noqa E501
-    persistence_layer = MockPersistenceLayer(expected_idempotency_key=idempotency_key)
-
-    class PaymentInput(BaseModel):
-        customer_id: str
-        transaction_id: str
-
-    class PaymentOutput(BaseModel):
-        customer_id: str
-        transaction_id: str
-
-    if output_serializer_type == "explicit":
-        output_serializer = PydanticSerializer(
-            model=PaymentOutput,
-        )
-    else:
-        output_serializer = PydanticSerializer
-
-    @idempotent_function(
-        data_keyword_argument="payment",
-        persistence_store=persistence_layer,
-        config=config,
-        output_serializer=output_serializer,
-    )
-    def collect_payment(payment: PaymentInput) -> PaymentOutput:
-        return PaymentOutput(**payment.dict())
-
-    # WHEN
-    payment = PaymentInput(**mock_event)
-    first_call: PaymentOutput = collect_payment(payment=payment)
-    assert first_call.customer_id == payment.customer_id
-    assert first_call.transaction_id == payment.transaction_id
-    assert isinstance(first_call, PaymentOutput)
-    second_call: PaymentOutput = collect_payment(payment=payment)
-    assert isinstance(second_call, PaymentOutput)
-    assert second_call.customer_id == payment.customer_id
-    assert second_call.transaction_id == payment.transaction_id
-
-
-def test_idempotent_function_serialization_pydantic_failure_no_return_type():
-    # GIVEN
-    config = IdempotencyConfig(use_local_cache=True)
-    mock_event = {"customer_id": "fake", "transaction_id": "fake-id"}
-    idempotency_key = f"{TESTS_MODULE_PREFIX}.test_idempotent_function_serialization_pydantic_failure_no_return_type.<locals>.collect_payment#{hash_idempotency_key(mock_event)}"  # noqa E501
-    persistence_layer = MockPersistenceLayer(expected_idempotency_key=idempotency_key)
-
-    class PaymentInput(BaseModel):
-        customer_id: str
-        transaction_id: str
-
-    class PaymentOutput(BaseModel):
-        customer_id: str
-        transaction_id: str
-
-    idempotent_function_decorator = idempotent_function(
-        data_keyword_argument="payment",
-        persistence_store=persistence_layer,
-        config=config,
-        output_serializer=PydanticSerializer,
-    )
-    with pytest.raises(IdempotencyNoSerializationModelError, match="No serialization model was supplied"):
-
-        @idempotent_function_decorator
-        def collect_payment(payment: PaymentInput):
-            return PaymentOutput(**payment.dict())
-
-
-def test_idempotent_function_serialization_pydantic_failure_bad_type():
-    # GIVEN
-    config = IdempotencyConfig(use_local_cache=True)
-    mock_event = {"customer_id": "fake", "transaction_id": "fake-id"}
-    idempotency_key = f"{TESTS_MODULE_PREFIX}.test_idempotent_function_serialization_pydantic_failure_no_return_type.<locals>.collect_payment#{hash_idempotency_key(mock_event)}"  # noqa E501
-    persistence_layer = MockPersistenceLayer(expected_idempotency_key=idempotency_key)
-
-    class PaymentInput(BaseModel):
-        customer_id: str
-        transaction_id: str
-
-    class PaymentOutput(BaseModel):
-        customer_id: str
-        transaction_id: str
-
-    idempotent_function_decorator = idempotent_function(
-        data_keyword_argument="payment",
-        persistence_store=persistence_layer,
-        config=config,
-        output_serializer=PydanticSerializer,
-    )
-    with pytest.raises(IdempotencyModelTypeError, match="Model type is not inherited from pydantic BaseModel"):
-
-        @idempotent_function_decorator
-        def collect_payment(payment: PaymentInput) -> dict:
-            return PaymentOutput(**payment.dict())
-
-
-@pytest.mark.parametrize("output_serializer_type", ["explicit", "deduced"])
 def test_idempotent_function_serialization_dataclass(output_serializer_type: str):
     # GIVEN
     dataclasses = get_dataclasses_lib()
@@ -1490,37 +1385,6 @@ def test_idempotent_function_serialization_dataclass_failure_no_return_type():
 
         @idempotent_function_decorator
         def collect_payment(payment: PaymentInput):
-            return PaymentOutput(**payment.dict())
-
-
-def test_idempotent_function_serialization_dataclass_failure_bad_type():
-    # GIVEN
-    dataclasses = get_dataclasses_lib()
-    config = IdempotencyConfig(use_local_cache=True)
-    mock_event = {"customer_id": "fake", "transaction_id": "fake-id"}
-    idempotency_key = f"{TESTS_MODULE_PREFIX}.test_idempotent_function_serialization_pydantic_failure_no_return_type.<locals>.collect_payment#{hash_idempotency_key(mock_event)}"  # noqa E501
-    persistence_layer = MockPersistenceLayer(expected_idempotency_key=idempotency_key)
-
-    @dataclasses.dataclass
-    class PaymentInput:
-        customer_id: str
-        transaction_id: str
-
-    @dataclasses.dataclass
-    class PaymentOutput:
-        customer_id: str
-        transaction_id: str
-
-    idempotent_function_decorator = idempotent_function(
-        data_keyword_argument="payment",
-        persistence_store=persistence_layer,
-        config=config,
-        output_serializer=PydanticSerializer,
-    )
-    with pytest.raises(IdempotencyModelTypeError, match="Model type is not inherited from pydantic BaseModel"):
-
-        @idempotent_function_decorator
-        def collect_payment(payment: PaymentInput) -> dict:
             return PaymentOutput(**payment.dict())
 
 
@@ -1804,24 +1668,6 @@ def test_idempotent_function_dataclasses():
     assert as_dict == expected_result
 
 
-def test_idempotent_function_pydantic():
-    # Scenario _prepare_data should convert a pydantic to a dict
-    class Foo(BaseModel):
-        name: str
-
-    expected_result = {"name": "Bar"}
-    data = Foo(name="Bar")
-    as_dict = _prepare_data(data)
-    assert as_dict == data.dict()
-    assert as_dict == expected_result
-
-
-@pytest.mark.parametrize("data", [None, "foo", ["foo"], 1, True, {}])
-def test_idempotent_function_other(data):
-    # All other data types should be left as is
-    assert _prepare_data(data) == data
-
-
 def test_idempotent_function_dataclass_with_jmespath():
     # GIVEN
     dataclasses = get_dataclasses_lib()
@@ -1832,29 +1678,6 @@ def test_idempotent_function_dataclass_with_jmespath():
 
     @dataclasses.dataclass
     class Payment:
-        customer_id: str
-        transaction_id: str
-
-    @idempotent_function(data_keyword_argument="payment", persistence_store=persistence_layer, config=config)
-    def collect_payment(payment: Payment):
-        return payment.transaction_id
-
-    # WHEN
-    payment = Payment(**mock_event)
-    result = collect_payment(payment=payment)
-
-    # THEN idempotency key assertion happens at MockPersistenceLayer
-    assert result == payment.transaction_id
-
-
-def test_idempotent_function_pydantic_with_jmespath():
-    # GIVEN
-    config = IdempotencyConfig(event_key_jmespath="transaction_id", use_local_cache=True)
-    mock_event = {"customer_id": "fake", "transaction_id": "fake-id"}
-    idempotency_key = f"{TESTS_MODULE_PREFIX}.test_idempotent_function_pydantic_with_jmespath.<locals>.collect_payment#{hash_idempotency_key(mock_event['transaction_id'])}"  # noqa E501
-    persistence_layer = MockPersistenceLayer(expected_idempotency_key=idempotency_key)
-
-    class Payment(BaseModel):
         customer_id: str
         transaction_id: str
